add - move moderation logic into util func/controller
